feat: separate code between node and web; add support for inference using web wallets
