+++ src/common/classes/model.ts
