+++ src/common/classes/operator.ts
