+++ src/common/classes/script.ts
