+++ src/common/queries/operator.ts
