+++ src/common/queries/script.ts
